requirements.txt
# Core dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langgraph>=0.0.30
llama-parse>=0.1.0
llama-index>=0.10.0
yfinance>=0.2.0
requests>=2.31.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# PDF processing
PyPDF2>=3.0.0

# Embeddings
sentence-transformers>=2.2.0

# dbt
dbt-postgres>=1.7.0

# Streamlit
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
